+++ demo/webpack/src/polyfills.ts
@@ -1,3 +1,0 @@
-import "core-js/es6";
-import "reflect-metadata";
-import "zone.js/dist/zone";