+++ src/angular2-toaster/src/public-api.ts
@@ -1,0 +1,12 @@
+/*
+ * Public API Surface of angular2-toaster
+ */
+
+export * from './lib/toast.component';
+export * from './lib/toaster-container.component';
+export * from './lib/toaster.service';
+export * from './lib/clearWrapper';
+export * from './lib/toaster-config';
+export * from './lib/toast';
+export * from './lib/bodyOutputType';
+export * from './lib/toaster.module';