+++ src/clearWrapper.ts
@@ -1,4 +1,0 @@
-export interface IClearWrapper {
-    toastId?: string;
-    toastContainerId?: number;
-}